Added user account feature, profile page

+++ app/utils/libs/supabaseClient.ts
@@ -1,13 +1,23 @@
 import { supabase } from "../supabase/client";
 
-export async function fetchCommunityPosts(limit: number) {
-  const { data, error } = await supabase
+export async function fetchCommunityPosts(
+  limit: number,
+  offset: number,
+  userId?: string,
+) {
+  let query = supabase
     .from("community_posts")
     .select(
-      `id, user_id, scenario, excuse, total_score, likes(count), users!community_posts_user_id_fkey ( username, avatar )`,
+      `id, user_id, scenario, excuse, total_score, likes(count), users!community_posts_user_id_fkey ( username, avatar ), created_at`,
     )
     .order("created_at", { ascending: false })
-    .limit(limit);
+    .range(offset, offset + limit - 1);
+
+  if (userId) {
+    query = query.eq("user_id", userId);
+  }
+
+  const { data, error } = await query;
 
   if (error || !data) {
     console.error(error);
@@ -53,6 +63,7 @@ export async function fetchCommunityPosts(limit: number) {
       score: post.total_score,
       likes: likeCount,
       isLiked: likedPostIds.includes(post.id),
+      created_at: post.created_at,
     };
   });
 }
@@ -110,11 +121,15 @@ export async function addComment(postId: number, comment: string) {
     return "You must be logged in to comment.";
   }
 
-  const { data, error } = await supabase.from("comments").insert({
-    user_id: user.id,
-    post_id: postId,
-    comment,
-  }).select("id").single();
+  const { data, error } = await supabase
+    .from("comments")
+    .insert({
+      user_id: user.id,
+      post_id: postId,
+      comment,
+    })
+    .select("id")
+    .single();
 
   if (error) {
     console.error(error);
@@ -137,6 +152,59 @@ export async function deleteComment(commentId: number) {
     .from("comments")
     .delete()
     .eq("id", commentId)
+    .eq("user_id", user.id);
+
+  if (error) {
+    console.error(error);
+    return error.message;
+  }
+}
+
+export async function deleteUserProfile(user_id: string) {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  if (!user) {
+    console.error("You must be logged in to delete your profile.");
+    return "You must be logged in to delete your profile.";
+  }
+
+  if (user.id !== user_id) {
+    console.error("You can only delete your own profile.");
+    return "You can only delete your own profile.";
+  }
+
+  const { error } = await supabase
+    .from("users")
+    .delete()
+    .eq("user_id", user_id);
+
+  if (error) {
+    console.error(error);
+    return error.message;
+  }
+}
+
+export async function updateUserProfile(
+  username: string | undefined,
+  avatarUrl?: string | undefined,
+) {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  if (!user) {
+    console.error("You must be logged in to update your profile.");
+    return "You must be logged in to update your profile.";
+  }
+
+  const { error } = await supabase
+    .from("users")
+    .update({
+      username: username,
+      avatar: avatarUrl,
+    })
     .eq("user_id", user.id);
 
   if (error) {
